one database; kivy gui is entry
